[Settings] Avoid from immediate update when UI inactive and SIM absent
When SIM absent during UI inactive, update can be postponed under the
condition of Activity onPause() or onStop(). And update when onResume()
or onStart().

This is not simply an improvement on performance but also a better user
experience.

User might unplug a SIM and plug-in again, and there're cases where a
SIM status might change into absent and back to active again due to some
mode switch in RIL/modem. From user perspective, the SIM status remain
the same and update of content is expected instead of closing the UI and
re-enter.

Bug: 195631787
Test: local
Change-Id: I8248e59895631dc90cf3831398e387b93483280c
(cherry picked from commit b2a6266fa91d66b84a540fc8eee891a83fda5538)